# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance backend API for HiDocs Form \u0026 Exam Maker application using Golang and Domain-Driven Design (DDD).",
        "title": "HiDocs Backend API (Form \u0026 Exam Maker)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "HiDocs Support",
            "email": "support@hidocs.id"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/forms/{form_id}/questions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Questions"
                ],
                "summary": "List questions for a form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "form_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.QuestionDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Questions"
                ],
                "summary": "Add a question to a form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "form_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Question Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.QuestionDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/options/{option_id}": {
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "Delete a question option",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Option ID",
                        "name": "option_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/questions/upload-image": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Questions"
                ],
                "summary": "Upload image attachment for questions",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image File (JPG, PNG, GIF)",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UploadImageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/questions/{question_id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Questions"
                ],
                "summary": "Update a question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Question ID",
                        "name": "question_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Question Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.QuestionDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "Delete a question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Question ID",
                        "name": "question_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "domain.QuestionType": {
            "type": "string",
            "enum": [
                "SHORT_TEXT",
                "LONG_TEXT",
                "MULTIPLE_CHOICE",
                "CHECKBOXES",
                "DROPDOWN",
                "RATING",
                "YES_NO",
                "MATH",
                "CODE",
                "IMAGE"
            ],
            "x-enum-varnames": [
                "TypeShortText",
                "TypeLongText",
                "TypeMultipleChoice",
                "TypeCheckboxes",
                "TypeDropdown",
                "TypeRating",
                "TypeYesNo",
                "TypeMath",
                "TypeCode",
                "TypeImage"
            ]
        },
...
        "dto.CreateOptionRequest": {
            "type": "object",
            "required": [
                "option_text"
            ],
            "properties": {
                "is_correct": {
                    "type": "boolean"
                },
                "option_text": {
                    "type": "string"
                },
                "order_index": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateQuestionRequest": {
            "type": "object",
            "required": [
                "question_text",
                "question_type"
            ],
            "properties": {
                "code_language": {
                    "type": "string"
                },
                "img_url": {
                    "type": "string"
                },
                "is_auto_scored": {
                    "type": "boolean"
                },
                "is_required": {
                    "type": "boolean"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateOptionRequest"
                    }
                },
                "order_index": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer",
                    "minimum": 0
                },
                "question_text": {
                    "type": "string"
                },
                "question_type": {
                    "$ref": "#/definitions/domain.QuestionType"
                }
            }
        },
...
        "dto.OptionDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "option_text": {
                    "type": "string"
                },
                "order_index": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.QuestionDTO": {
            "type": "object",
            "properties": {
                "code_language": {
                    "type": "string"
                },
                "form_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "img_url": {
                    "type": "string"
                },
                "is_auto_scored": {
                    "type": "boolean"
                },
                "is_required": {
                    "type": "boolean"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OptionDTO"
                    }
                },
                "order_index": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "question_text": {
                    "type": "string"
                },
                "question_type": {
                    "$ref": "#/definitions/domain.QuestionType"
                }
            }
        },
...
        "dto.UpdateQuestionRequest": {
            "type": "object",
            "required": [
                "question_text",
                "question_type"
            ],
            "properties": {
                "code_language": {
                    "type": "string"
                },
                "img_url": {
                    "type": "string"
                },
                "is_auto_scored": {
                    "type": "boolean"
                },
                "is_required": {
                    "type": "boolean"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateOptionRequest"
                    }
                },
                "order_index": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer",
                    "minimum": 0
                },
                "question_text": {
                    "type": "string"
                },
                "question_type": {
                    "$ref": "#/definitions/domain.QuestionType"
                }
            }
        },
        "dto.UploadImageResponse": {
            "type": "object",
            "properties": {
                "img_url": {
                    "type": "string"
                }
            }
        },
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}